+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * [services] in stats, if user unknown ie not found, show *unknown* value
 * [core] remove CoreStatus sql related calls, use static values
 * [core] get full user info on getUserByLogin, use id instead of idUser field
+* [core] allow password reset using login
 
 ## 2.9.4