+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes
 
+## 2.10.0
+
+* [core][cli] manage email checks and account inactivation (at space level)
+  if user does not validate his email
+
 ## 2.9.5
 
 * code cleanup
@@ -11,8 +16,8 @@
 * [core] set minimal level for module access #511
 * [services] in stats, if user unknown ie not found, show *unknown* value
 * [core] remove CoreStatus sql related calls, use static values
-* [core][cli] manage email checks and account inactivation (at space level)
-  if user does not validate his email
+* [core] get full user info on getUserByLogin, use id instead of idUser field
+* [core] allow password reset using login
 
 ## 2.9.4